Added Errors and new Template

+++ src/routes/SyntaxDefinitionRoutes.ts
@@ -1,6 +1,7 @@
 import { NextFunction } from "connect";
 import { Request, Response, Router } from "express";
 import * as FileUtil from "../util/FileUtil";
+import * as ErrorUtil from "../util/ErrorUtil";
 
 class SyntaxDefinitionRoutes {
 
@@ -15,7 +16,9 @@ class SyntaxDefinitionRoutes {
 
   private setup() {
     this.router.param("name", (req: Request, _res: Response, next: NextFunction, name: string) => {
-      req.fields.name = name;
+      if (req.fields) {
+        req.fields.name = name;
+      }
       next();
     });
     this.router.get("/", (_req: Request, res: Response, next: NextFunction) => {
@@ -24,12 +27,16 @@ class SyntaxDefinitionRoutes {
         .catch(err => next(err));
     });
     this.router.get("/:name", (req: Request, res: Response, next: NextFunction) => {
-      let file = <string>req.fields.name;
-      FileUtil.read(this.folder, file)
-      .then(buffer => {
-        res.type("xml");
-        res.send(buffer.toString("utf8"));
-      }).catch(err => next(err));
+      if (req.fields && req.fields.name) {
+        let file = <string>req.fields.name;
+        FileUtil.read(this.folder, file)
+        .then(buffer => {
+          res.type("xml");
+          res.send(buffer.toString("utf8"));
+        }).catch(err => next(err));
+      } else {
+        next(ErrorUtil.MissingFieldError("name"));
+      }
     });
     this.router.post("/", (req: Request, res: Response, next: NextFunction) => {
       FileUtil.saveFiles(req, this.folder)
@@ -37,17 +44,22 @@ class SyntaxDefinitionRoutes {
       .catch(err => next(err));
     });
     this.router.delete("/:name", (req: Request, res: Response, next: NextFunction) => {
-      let name = <string>req.fields.name;
-      FileUtil.deleteDir(this.folder, name)
-      .then(() => res.end())
-      .catch(err => next(err));
+      if (req.fields && req.fields.name) {
+        let name = <string>req.fields.name;
+        FileUtil.deleteDir(this.folder, name)
+        .then(() => res.end())
+        .catch(err => next(err));
+      }
+      else {
+        next(ErrorUtil.MissingFieldError("name"));
+      }
     });
   }
 
   private async getFiles(): Promise<string[]> {
-    let stats = await FileUtil.readdirStats(this.folder)
+    let stats = FileUtil.readdirStats(this.folder);
     let definitions: string[] = [];
-    for (let stat of stats) {
+    for await (let stat of stats) {
       let name = stat[0];
       let file = stat[1];
       if (file.isFile() && name.endsWith(".xml")) {
