+++ src/routes/TemplateRoutes.ts
@@ -2,6 +2,7 @@ import { NextFunction } from "connect";
 import { Request, Response, Router } from "express";
 import { Template } from "../model/Template";
 import * as FileUtil from "../util/FileUtil";
+import * as ErrorUtil from "../util/ErrorUtil";
 
 class TemplateRoutes {
 
@@ -14,7 +15,9 @@ class TemplateRoutes {
 
   private setup() {
     this.router.param("name", (req: Request, _res: Response, next: NextFunction, name: string) => {
-      req.fields.name = name;
+      if (req.fields) {
+        req.fields.name = name;
+      }
       next();
     });
     this.router.get("/", (_req: Request, res: Response, next: NextFunction) => {
@@ -32,41 +35,31 @@ class TemplateRoutes {
       .catch(err => next(err));
     });
     this.router.delete("/:name", (req: Request, res: Response, next: NextFunction) => {
-      let name = <string>req.fields.name;
-      FileUtil.deleteDir(FileUtil.resources, FileUtil.templates, name)
-      .then(() => res.end())
-      .catch(err => next(err));
+      if (req.fields && req.fields.name) {
+        let name = <string>req.fields.name;
+        FileUtil.deleteDir(FileUtil.resources, FileUtil.templates, name)
+        .then(() => res.end())
+        .catch(err => next(err));
+      } else {
+        next(ErrorUtil.MissingFieldError("name"));
+      }
     });
   }
 
   private async getDirectories(): Promise<Template[]> {
-    let folder = FileUtil.combine(process.cwd(), FileUtil.resources, FileUtil.templates);
-    let stats = await FileUtil.readdirStats(folder);
+    let stats = FileUtil.readdirStats(FileUtil.resources, FileUtil.templates);
     let templates: Template[] = [];
-    for (let stat of stats) {
+    for await (let stat of stats) {
       if (stat[1].isDirectory()) {
         let temp = new Template();
-        temp.name = stat[0];
-        temp.icon = await this.getIcon(folder, temp.name);
+        await temp.load(FileUtil.resources, FileUtil.templates, stat[0]);
         templates.push(temp);
       }
     }
     return templates;
   }
 
-  private async getIcon(...folder: string[]): Promise<string> {
-    let svgFile = FileUtil.combine(...folder, "icon.svg");
-    try
-    {
-      let buffer = await FileUtil.read(svgFile);
-      let svg = buffer.toString("utf8");
-      return svg;
-    }
-    catch
-    {
-      return undefined;
-    }
-  }
+  
 }
 
 export default new TemplateRoutes().router;
